requirements.txt
# ============================================================================
# AML Therapeutic Target Identification Pipeline - Requirements
# ============================================================================
# Python 3.9+ recommended
#
# Installation: pip install -r requirements.txt
#
# Note: For GPU acceleration (scVI/scANVI training), install PyTorch with
# appropriate CUDA/MPS support first:
# - macOS (Apple Silicon): pip install torch torchvision (MPS enabled)
# - Linux/Windows (NVIDIA): See https://pytorch.org/get-started/locally/
# ============================================================================

# Core scientific computing
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.9.0

# Single-cell analysis
scanpy>=1.9.0
anndata>=0.9.0
scvi-tools>=1.0.0
celltypist>=1.6.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
torch>=2.0.0

# Differential expression
pydeseq2>=0.4.0
statsmodels>=0.14.0

# Pathway enrichment
gseapy>=1.0.0

# Dimensionality reduction
umap-learn>=0.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
openpyxl>=3.1.0

# Optional but recommended
jupyter>=1.0.0
ipykernel>=6.0.0
